+++ actix-test/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Minimum supported Rust version (MSRV) is now 1.72.
+- TestServer can now be told to listen on an interface other than `localhost`, via the `TestServerConfig::listen_address()` method.
 
 ## 0.1.3